Refactor index.js file for each main object.
Extract requestAnimationFrame logic to separate functions.
