+++ lib/helicopter.js
@@ -1,0 +1,9 @@
+function Helicopter(x, y, width, height, isFlying) {
+  this.x = x;
+  this.y = y;
+  this.width = width;
+  this.height = height;
+  this.isFlying = isFlying;
+}
+
+module.exports = Helicopter
