+++ lib/index.js
@@ -1,21 +1,6 @@
-function World(gravatationalVelocity) {
-  this.gravatationalVelocity = gravatationalVelocity;
-}
-// TODO: refactor: make a base class called block to use for helicopter and obstacle
-function Helicopter(x, y, width, height, isFlying) {
-  this.x = x;
-  this.y = y;
-  this.width = width;
-  this.height = height;
-  this.isFlying = isFlying;
-}
-
-function Obstacle(x, y, width, height) {
-  this.x = x;
-  this.y = y;
-  this.width = width;
-  this.height = height;
-}
+var World = require('./world')
+var Helicopter = require('./helicopter')
+var Obstacle = require('./obstacle')
 
 var canvas = document.getElementById('game');
 var context = canvas.getContext('2d');
@@ -31,34 +16,48 @@ var helicopterBottomLimit = canvas.height - helicopter.height;
 window.requestAnimationFrame(function gameLoop() {
   context.clearRect(0, 0, canvas.width, canvas.height);
 
-  if (helicopter.isFlying) {
-    var nextBlockY = helicopter.y - world.gravatationalVelocity;
-    helicopter.y = Math.max(nextBlockY, helicopterTopLimit);
-  }
-  else {
-    var nextBlockY = helicopter.y + world.gravatationalVelocity;
-    helicopter.y = Math.min(nextBlockY, helicopterBottomLimit);
-  }
+  checkFlight();
 
-  // move obstacle1 from right to left
   obstacle1.x -= 4
-  if (obstacle1.x < -obstacle1.width) {
-    obstacle1.x = canvas.width
-  }
+  moveObstacle(obstacle1);
 
   obstacle2.x -= 3
-  if (obstacle2.x < -obstacle2.width) {
-    obstacle2.x = canvas.width
-  }
-  context.fillRect(helicopter.x, helicopter.y, helicopter.width, helicopter.height);
-  context.fillRect(obstacle1.x, obstacle1.y, obstacle1.width, obstacle1.height);
-  context.fillRect(obstacle2.x, obstacle2.y, obstacle2.width, obstacle2.height);
+  moveObstacle(obstacle2);
+
+  drawHelicopter();
+  drawObstacles();
 
   window.requestAnimationFrame(gameLoop);
 });
 
 window.addEventListener('keydown', startFlying);
 window.addEventListener('keyup', stopFlying);
+
+function drawObstacles() {
+  context.fillRect(obstacle1.x, obstacle1.y, obstacle1.width, obstacle1.height);
+  context.fillRect(obstacle2.x, obstacle2.y, obstacle2.width, obstacle2.height);
+}
+
+function drawHelicopter(){
+  context.fillRect(helicopter.x, helicopter.y, helicopter.width, helicopter.height);
+}
+
+function moveObstacle(obstacle){
+  if (obstacle.x < -obstacle.width) {
+    obstacle.x = canvas.width
+  }
+}
+
+function checkFlight(){
+  if (helicopter.isFlying) {
+    var nextBlockY = helicopter.y - 3;
+    helicopter.y = Math.max(nextBlockY, helicopterTopLimit);
+  }
+  else {
+    var nextBlockY = helicopter.y + 3;
+    helicopter.y = Math.min(nextBlockY, helicopterBottomLimit);
+  }
+}
 
 function startFlying(e) {
   if (e.keyCode === 85) {
