+++ lib/obstacle.js
@@ -1,0 +1,8 @@
+function Obstacle(x, y, width, height) {
+  this.x = x;
+  this.y = y;
+  this.width = width;
+  this.height = height;
+}
+
+module.exports = Obstacle
