+++ lib/world.js
@@ -1,0 +1,5 @@
+function World(gravatationalVelocity) {
+  this.gravatationalVelocity = gravatationalVelocity;
+}
+
+module.exports = World
